Separate uniqueness of SectorConfig views into its own class that inherents from the default Views class.
